L.MapをCreationMapViewコンポーネントとして分離
