+++ src/components/CreationMapView/index.vue
@@ -1,0 +1,24 @@
+<template>
+    <div id="creation-map-view">
+      <div id="map"></div>
+    </div>
+</template>
+
+<script lang='ts' src='./index.ts'/>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped>
+html,
+body,
+#creation-map-view {
+  height: 100%;
+}
+#map {
+  height: 100%;
+}
+
+body {
+  margin: 0;
+  height: 100%;
+}
+</style>
